agent/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
httpx
groq
python-dotenv
